distributed_troubleshooter/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
spacy
python-jose
bcrypt
uvicorn[standard]
sentence-transformers